responding to a round of comments
